django-backend/requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
python-dotenv==0.19.2
supertokens-python
django-extensions
djangorestframework==3.14.0
